+++ controllers/notices/listFavorite.js
@@ -1,6 +1,24 @@
 const Notice = require("../../models/notice");
 const User = require("../../models/user");
+const { ageNotice } = require("../../helpers");
+
 const listFavorite = async (req, res) => {
-  res.json({ ok: true });
+  const { _id } = req.user;
+  const { page = 1, limit = 12, query = "", age, sex } = req.query;
+  const skip = (page - 1) * limit;
+  const user = await User.findById(_id);
+  let notices = await Notice.find({$and:[{ _id: { $in: user.favoritesArr } }, { title: { $regex: query } }]}, "", {
+    limit,
+    skip,
+  });
+  if (age) {
+    notices = notices.filter((notice) => ageNotice(notice, age));
+  }
+
+  if (sex) {
+    notices = notices.filter((notice) => notice.sex === sex);
+  }
+
+  res.json(notices);
 };
 module.exports = listFavorite;
