+++ controllers/notices/listLostFound.js
@@ -1,10 +1,10 @@
 const { filterNotices } = require("../../helpers");
 
 const listLostFound = async (req, res) => {
-  const { query = "", age, sex } = req.query;
+  const { query = "", age, sex, page, limit } = req.query;
   const category = "lost-found";
 
-  const notices = await filterNotices(category, query, age, sex);
+  const notices = await filterNotices(category, query, age, sex, page, limit);
   res.json({ notices });
 };
