+++ helpers/index.js
@@ -1,9 +1,11 @@
 const controllerWrapper = require("./controllerWrapper");
 const requestError = require("./requestError");
 const filterNotices = require("./filterNotices");
+const ageNotice = require("./ageNotice")
 
 module.exports = {
   controllerWrapper,
   requestError,
   filterNotices,
+  ageNotice,
 };
